requirements.txt
appnope==0.1.3
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
black==22.3.0
certifi==2021.10.8
click==8.1.2
cycler==0.11.0
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fonttools==4.31.2
ipython==8.2.0
jedi @ file:///opt/concourse/worker/volumes/live/fb2e45f1-1dff-42fa-6cab-bd094a4c4372/volume/jedi_1644315264796/work
kiwisolver==1.4.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
numpy==1.22.3
ola==0.10.7
opencv-contrib-python==4.5.5.64
packaging==21.3
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathspec==0.9.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.0.1
platformdirs==2.5.1
prompt-toolkit==3.0.29
protobuf==3.20.0
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
PyOpenGL==3.1.6
PyOpenGL-accelerate==3.1.5
pyparsing==3.0.7
python-dateutil==2.8.2
python-osc==1.8.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
tomli==2.0.1
tqdm==4.63.1
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing_extensions==4.1.1
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
